% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_paths.R
\name{aggregate_paths}
\alias{aggregate_paths}
\title{Aggregate the paths through a random forest}
\usage{
aggregate_paths(data, rf)
}
\arguments{
\item{data}{the output of \code{\link{rf_all_paths}}.}

\item{rf}{an object of class \code{randomForest} retuned by \code{\link[randomForest]{randomForest}}. This is the random forest through which the paths described in \code{data} were calculated i.e. that which was supplied to \code{\link{rf_all_paths}}.}
}
\description{
\code{\link{aggregate_paths}} is a low level function called by \code{\link{rf_pathfinder}} that aggregates all the paths through a random forest ready for visualisation as a parallel coorindate plot with \code{\link{rf_parcoor}} or a Sankey diagram with \code{\link{rf_sankey}}.
}
