% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_all_paths.R
\name{rf_all_paths}
\alias{rf_all_paths}
\title{All paths through a random forest}
\usage{
rf_all_paths(rf, report.progress = TRUE)
}
\arguments{
\item{rf}{An object of class \code{randomForest} returned by \code{\link[randomForest]{randomForest}}}

\item{report.progress}{if \code{TRUE} the percentage of trees for which all paths have been found is printed to the console as the function runs, if \code{FALSE} no such progress report is printed}
}
\description{
\code{rf_all_paths} is a low level function called by \code{\link{rf_pathfinder}} to determine all paths through all trees in a random forest.
}
