% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_pathfinder.R
\name{rf_pathfinder}
\alias{rf_pathfinder}
\title{Find all possible paths through a random forest}
\usage{
rf_pathfinder(rf, report.progress = TRUE)
}
\arguments{
\item{rf}{An object of class \code{randomForest} returned by \code{\link[randomForest]{randomForest}}}
}
\description{
\code{rf_pathfinder} is a high level function to find all paths through a random forest.  This function returns a dataframe that can be supplied to \code{\link{rf_parcoor}} or \code{\link{rf_sankey}} to produce plots of these paths. Please see the help pages for \code{\link{rf_parcoor}} or \code{\link{rf_sankey}} for examples of workflows flow for visualising the paths through random forests with these functions.
}
\seealso{
\code{\link{rf_parcoor}} and \code{\link{rf_sankey}}
}
