% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_paths.R
\name{tree_paths}
\alias{tree_paths}
\title{All paths through a single tree from a random forest}
\usage{
tree_paths(rf, tree, plot.network = FALSE, plot.vert.size = 0,
  label.vert.with.name = TRUE, circular.layout = TRUE,
  vertex.label.size = 0.75, arrow.size = 0.25)
}
\arguments{
\item{rf}{an object of class \code{randomForest} returned by \code{\link[randomForest]{randomForest}}}

\item{tree}{a number indexing the tree constituent to \code{rf} which will be extracted and converted to an \code{igraph} network}

\item{plot.network}{if \code{TRUE} the network is plotted. If \code{FALSE} the network is not plotted}

\item{plot.vert.size}{a number setting the size of the circles which represent the nodes in the plot of the network}

\item{label.vert.with.name}{if \code{TRUE} each vertex (node) is labeled with the covariate that defined the binary partition represented by that vertex. If \code{FALSE} each vertex is labeled with the name assigned to it by \code{igraph}}

\item{circular.layout}{if \code{TRUE} the network is plotted with the circular layout from the igraph package. If \code{FALSE} the network is plotted with the default layout from the \code{igraph} package}

\item{vertex.label.size}{a number controlling the size of the text used for the vertex labels in the plot}

\item{edge.arrow.size}{a number controlling the size of arrows that represent the graph edges in the plot}
}
\description{
\code{tree_paths} is a low level function called by \code{\link{rf_all_paths}} to find all paths from root to leaf nodes through a single tree extracted from a random forest.
}
\seealso{
\code{\link{rf_all_paths}}
}
